# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateScanRequest:
    properties:
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      regions:
        example:
        - us-east-1
        - eu-west-1
        items:
          type: string
        minItems: 1
        type: array
      resource_types:
        example:
        - ec2_instance
        - ebs_volume
        items:
          type: string
        type: array
    required:
    - organization_id
    - provider
    - regions
    type: object
  handler.CreateScanResponse:
    properties:
      data:
        $ref: '#/definitions/handler.ScanDTO'
      message:
        example: scan created and queued for processing
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  handler.PaginatedResponse:
    properties:
      data: {}
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  handler.ResourceDTO:
    properties:
      carbon_footprint_kg:
        example: 12.5
        type: number
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_seen_at:
        type: string
      monthly_cost:
        example: 45.5
        type: number
      name:
        example: my-instance
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      region:
        example: us-east-1
        type: string
      resource_id:
        example: i-1234567890abcdef0
        type: string
      status:
        enum:
        - active
        - unused
        - deleted
        - excluded
        example: unused
        type: string
      tags:
        additionalProperties:
          type: string
        type: object
      type:
        example: ec2_instance
        type: string
      updated_at:
        type: string
    type: object
...
  handler.ScanDTO:
    properties:
      carbon_savings_kg:
        example: 45.5
        type: number
      completed_at:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      estimated_savings:
        example: 1250
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      organization_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      provider:
        enum:
        - aws
        - azure
        - gcp
        example: aws
        type: string
      regions:
        example:
        - us-east-1
        - eu-west-1
        items:
          type: string
        type: array
      resource_types:
        example:
        - ec2_instance
        - ebs_volume
        items:
          type: string
        type: array
      resources_found:
        example: 150
        type: integer
      started_at:
        type: string
      status:
        enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
        example: completed
        type: string
      unused_found:
        example: 23
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cloudsweep.io
    name: CloudSweep Support
    url: https://github.com/ermt0504/cloud-sweep
  description: API FinOps/GreenOps pour automatiser le nettoyage des ressources cloud
    inutilisees
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CloudSweep API
  version: "1.0"
paths:
...
  /resources:
    get:
      consumes:
      - application/json
      description: Get a paginated list of cloud resources with optional filters
      parameters:
      - description: Filter by cloud provider
        enum:
        - aws
        - azure
        - gcp
        in: query
        name: provider
        type: string
      - description: Filter by resource type
        in: query
        name: type
        type: string
      - description: Filter by status
        enum:
        - active
        - unused
        - deleted
        - excluded
        in: query
        name: status
        type: string
      - description: Filter by region
        in: query
        name: region
        type: string
      - default: 50
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.ResourceDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List resources
      tags:
      - Resources
  /resources/{id}:
    delete:
      consumes:
      - application/json
      description: Mark a resource as deleted
      parameters:
      - description: Resource ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete resource
      tags:
      - Resources
    get:
      consumes:
      - application/json
      description: Get a single cloud resource by its ID
      parameters:
      - description: Resource ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.ResourceDTO'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get resource by ID
      tags:
      - Resources
  /scans:
    get:
      consumes:
      - application/json
      description: Get a paginated list of scans with optional filters
      parameters:
      - description: Filter by cloud provider
        enum:
        - aws
        - azure
        - gcp
        in: query
        name: provider
        type: string
      - description: Filter by status
        enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
        in: query
        name: status
        type: string
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.ScanDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List scans
      tags:
      - Scans
    post:
      consumes:
      - application/json
      description: Create a new cloud resource scan and queue it for processing
      parameters:
      - description: Scan request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateScanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateScanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new scan
      tags:
      - Scans
  /scans/{id}:
    get:
      consumes:
      - application/json
      description: Get a single scan by its ID
      parameters:
      - description: Scan ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/handler.ScanDTO'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get scan by ID
      tags:
      - Scans
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"